Move PaddingNode to the correct namespace
