+++ SAGESharp.Tests/SAGESharp.Tests.csproj
@@ -88,12 +88,12 @@
     <Compile Include="IO\Binary\TreeBasedSerialization\EdgeTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\ListNodeTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\OffsetNodeTests.cs" />
+    <Compile Include="IO\Binary\TreeBasedSerialization\PaddingNodeTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeBinarySerializerTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeBuilderTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeReaderTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeWriterTests.cs" />
     <Compile Include="IO\Binary\WriterTests.cs" />
-    <Compile Include="IO\PaddingNodeTests.cs" />
     <Compile Include="IO\PrimitiveTypeDataNodeTests.cs" />
     <Compile Include="IO\StringDataNodeTests.cs" />
     <Compile Include="IO\Substitutes.cs" />
