+++ SAGESharp/IO/TreeBasedSerialization.cs
@@ -416,44 +416,5 @@ namespace SAGESharp.IO
 
         private static bool IsType(object value) => typeof(T) == value.GetType();
     }
-
-    internal sealed class PaddingNode : IDataNode
-    {
-        private readonly byte[] padding;
-
-        private readonly IDataNode childNode;
-
-        public PaddingNode(byte size, IDataNode childNode)
-        {
-            Validate.Argument(size > 0, "Padding size cannot be 0.");
-            Validate.ArgumentNotNull(childNode, nameof(childNode));
-
-            padding = new byte[size];
-            this.childNode = childNode;
-        }
-
-        public IReadOnlyList<IEdge> Edges { get => childNode.Edges; }
-
-        public object Read(IBinaryReader binaryReader)
-        {
-            Validate.ArgumentNotNull(binaryReader, nameof(binaryReader));
-
-            object result = childNode.Read(binaryReader);
-
-            binaryReader.ReadBytes(padding.Length);
-
-            return result;
-        }
-
-        public void Write(IBinaryWriter binaryWriter, object value)
-        {
-            Validate.ArgumentNotNull(binaryWriter, nameof(binaryWriter));
-            Validate.ArgumentNotNull(value, nameof(value));
-
-            childNode.Write(binaryWriter, value);
-
-            binaryWriter.WriteBytes(padding);
-        }
-    }
     #endregion
 }
